requirements.txt
requests
pandas
biopython
numpy
matplotlib
seaborn
jupyter
